rename referenceActions to accumulateActions
