+++ src/actions/index.js
@@ -1,3 +1,4 @@
+export * from './accumulateActions';
 export * from './authActions';
 export * from './dataActions';
 export * from './fetchActions';
@@ -5,5 +6,4 @@ export * from './filterActions';
 export * from './listActions';
 export * from './localeActions';
 export * from './notificationActions';
-export * from './referenceActions';
 export * from './uiActions';
